Packagify multiprocess bits
