+++ tests/test_multiprocess.py
@@ -6,14 +6,15 @@ import shutil
 import sys
 import tempfile
 
-from prometheus_client import core, mmap_dict, values
+from prometheus_client import core, values
 from prometheus_client.core import (
     CollectorRegistry, Counter, Gauge, Histogram, Sample, Summary,
 )
 from prometheus_client.multiprocess import (
     mark_process_dead, MultiProcessCollector,
-)
-from prometheus_client.values import MultiProcessValue, MutexValue
+    mmap_dict)
+from prometheus_client.values import MutexValue
+from prometheus_client.multiprocess.mmaped_value import MmapedValue
 
 if sys.version_info < (2, 7):
     # We need the skip decorators from unittest2 on Python 2.6.
@@ -28,7 +29,7 @@ class TestMultiProcess(unittest.TestCase):
     def setUp(self):
         self.tempdir = tempfile.mkdtemp()
         os.environ['prometheus_multiproc_dir'] = self.tempdir
-        values.ValueClass = MultiProcessValue(lambda: 123)
+        values.ValueClass = MmapedValue(lambda: 123)
         self.registry = CollectorRegistry()
         self.collector = MultiProcessCollector(self.registry, self.tempdir)
 
@@ -43,7 +44,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_counter_adds(self):
         c1 = Counter('c', 'help', registry=None)
-        values.ValueClass = MultiProcessValue(lambda: 456)
+        values.ValueClass = MmapedValue(lambda: 456)
         c2 = Counter('c', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('c_total'))
         c1.inc(1)
@@ -52,7 +53,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_summary_adds(self):
         s1 = Summary('s', 'help', registry=None)
-        values.ValueClass = MultiProcessValue(lambda: 456)
+        values.ValueClass = MmapedValue(lambda: 456)
         s2 = Summary('s', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('s_count'))
         self.assertEqual(0, self.registry.get_sample_value('s_sum'))
@@ -63,7 +64,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_histogram_adds(self):
         h1 = Histogram('h', 'help', registry=None)
-        values.ValueClass = MultiProcessValue(lambda: 456)
+        values.ValueClass = MmapedValue(lambda: 456)
         h2 = Histogram('h', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('h_count'))
         self.assertEqual(0, self.registry.get_sample_value('h_sum'))
@@ -76,7 +77,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_all(self):
         g1 = Gauge('g', 'help', registry=None)
-        values.ValueClass = MultiProcessValue(lambda: 456)
+        values.ValueClass = MmapedValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '123'}))
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '456'}))
@@ -88,7 +89,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_liveall(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='liveall')
-        values.ValueClass = MultiProcessValue(lambda: 456)
+        values.ValueClass = MmapedValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='liveall')
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '123'}))
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '456'}))
@@ -102,7 +103,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_min(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='min')
-        values.ValueClass = MultiProcessValue(lambda: 456)
+        values.ValueClass = MmapedValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='min')
         self.assertEqual(0, self.registry.get_sample_value('g'))
         g1.set(1)
@@ -111,7 +112,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_max(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='max')
-        values.ValueClass = MultiProcessValue(lambda: 456)
+        values.ValueClass = MmapedValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='max')
         self.assertEqual(0, self.registry.get_sample_value('g'))
         g1.set(1)
@@ -120,7 +121,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_livesum(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='livesum')
-        values.ValueClass = MultiProcessValue(lambda: 456)
+        values.ValueClass = MmapedValue(lambda: 456)
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='livesum')
         self.assertEqual(0, self.registry.get_sample_value('g'))
         g1.set(1)
@@ -136,7 +137,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_counter_across_forks(self):
         pid = 0
-        values.ValueClass = MultiProcessValue(lambda: pid)
+        values.ValueClass = MmapedValue(lambda: pid)
         c1 = Counter('c', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('c_total'))
         c1.inc(1)
@@ -148,7 +149,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_initialization_detects_pid_change(self):
         pid = 0
-        values.ValueClass = MultiProcessValue(lambda: pid)
+        values.ValueClass = MmapedValue(lambda: pid)
 
         # can not inspect the files cache directly, as it's a closure, so we
         # check for the actual files themselves
@@ -169,7 +170,7 @@ class TestMultiProcess(unittest.TestCase):
     @unittest.skipIf(sys.version_info < (2, 7), "Test requires Python 2.7+.")
     def test_collect(self):
         pid = 0
-        values.ValueClass = MultiProcessValue(lambda: pid)
+        values.ValueClass = MmapedValue(lambda: pid)
         labels = dict((i, i) for i in 'abcd')
 
         def add_label(key, value):
@@ -230,7 +231,7 @@ class TestMultiProcess(unittest.TestCase):
     @unittest.skipIf(sys.version_info < (2, 7), "Test requires Python 2.7+.")
     def test_merge_no_accumulate(self):
         pid = 0
-        values.ValueClass = MultiProcessValue(lambda: pid)
+        values.ValueClass = MmapedValue(lambda: pid)
         labels = dict((i, i) for i in 'abcd')
 
         def add_label(key, value):
